LPS-13086 Test Formatting
