+++ portal-web/test/com/liferay/portalweb/portlet/webcontentdisplay/wcstructure/WCStructureTestPlan.java
@@ -12,14 +12,10 @@
  * details.
  */
 
-package com.liferay.portalweb.portlet.webcontentdisplay;
+package com.liferay.portalweb.portlet.webcontentdisplay.wcstructure;
 
 import com.liferay.portalweb.portal.BaseTestSuite;
-import com.liferay.portalweb.portlet.webcontentdisplay.comment.CommentTestPlan;
-import com.liferay.portalweb.portlet.webcontentdisplay.portlet.PortletTestPlan;
-import com.liferay.portalweb.portlet.webcontentdisplay.wcstructure.WCStructureTestPlan;
-import com.liferay.portalweb.portlet.webcontentdisplay.wcwebcontent.WCWebContentTestPlan;
-import com.liferay.portalweb.portlet.webcontentdisplay.webcontent.WebContentTestPlan;
+import com.liferay.portalweb.portlet.webcontentdisplay.wcstructure.addwcstructure2parentstructurestructure1wcd.AddWCStructure2ParentStructureStructure1WCDTests;
 
 import junit.framework.Test;
 import junit.framework.TestSuite;
@@ -27,16 +23,13 @@ import junit.framework.TestSuite;
 /**
  * @author Brian Wing Shun Chan
  */
-public class WebContentDisplayTestPlan extends BaseTestSuite {
+public class WCStructureTestPlan extends BaseTestSuite {
 
 	public static Test suite() {
 		TestSuite testSuite = new TestSuite();
 
-		testSuite.addTest(CommentTestPlan.suite());
-		testSuite.addTest(PortletTestPlan.suite());
-		testSuite.addTest(WCStructureTestPlan.suite());
-		testSuite.addTest(WCWebContentTestPlan.suite());
-		testSuite.addTest(WebContentTestPlan.suite());
+		testSuite.addTest(
+			AddWCStructure2ParentStructureStructure1WCDTests.suite());
 
 		return testSuite;
 	}
