+++ portal-web/test/com/liferay/portalweb/portlet/webcontentdisplay/wcstructure/addwcstructure2parentstructurestructure1wcd/AddWCStructure1WCDTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portalweb.portlet.webcontentdisplay.webcontent.addwcstructure2parentstructure1wcd;
+package com.liferay.portalweb.portlet.webcontentdisplay.wcstructure.addwcstructure2parentstructurestructure1wcd;
 
 import com.liferay.portalweb.portal.BaseTestCase;
 import com.liferay.portalweb.portal.util.RuntimeVariables;
@@ -20,10 +20,8 @@ import com.liferay.portalweb.portal.util.RuntimeVariables;
 /**
  * @author Brian Wing Shun Chan
  */
-public class AddWCStructure2ParentStructureStructure1WCDTest
-	extends BaseTestCase {
-	public void testAddWCStructure2ParentStructureStructure1WCD()
-		throws Exception {
+public class AddWCStructure1WCDTest extends BaseTestCase {
+	public void testAddWCStructure1WCD() throws Exception {
 		selenium.open("/web/guest/home/");
 		selenium.clickAt("link=Web Content Display Test Page",
 			RuntimeVariables.replace("Web Content Display Test Page"));
@@ -105,127 +103,28 @@ public class AddWCStructure2ParentStructureStructure1WCDTest
 		selenium.clickAt("//input[@value='Add Structure']",
 			RuntimeVariables.replace("Add Structure"));
 		selenium.waitForPageToLoad("30000");
-
-		for (int second = 0;; second++) {
-			if (second >= 90) {
-				fail("timeout");
-			}
-
-			try {
-				if (selenium.isVisible("//input[@value='Select']")) {
-					break;
-				}
-			}
-			catch (Exception e) {
-			}
-
-			Thread.sleep(1000);
-		}
-
-		selenium.clickAt("//input[@value='Select']",
-			RuntimeVariables.replace("Select"));
-		selenium.selectFrame("relative=top");
-
-		for (int second = 0;; second++) {
-			if (second >= 90) {
-				fail("timeout");
-			}
-
-			try {
-				if (selenium.isVisible(
-							"//iframe[@name='_15_parentStructureSelector']")) {
-					break;
-				}
-			}
-			catch (Exception e) {
-			}
-
-			Thread.sleep(1000);
-		}
-
-		selenium.selectFrame("//iframe[@name='_15_parentStructureSelector']");
-
-		for (int second = 0;; second++) {
-			if (second >= 90) {
-				fail("timeout");
-			}
-
-			try {
-				if (selenium.isElementPresent(
-							"//script[contains(@src,'/liferay/navigation_interaction.js')]")) {
-					break;
-				}
-			}
-			catch (Exception e) {
-			}
-
-			Thread.sleep(1000);
-		}
-
-		for (int second = 0;; second++) {
-			if (second >= 90) {
-				fail("timeout");
-			}
-
-			try {
-				if (selenium.isVisible("//td[1]/a")) {
-					break;
-				}
-			}
-			catch (Exception e) {
-			}
-
-			Thread.sleep(1000);
-		}
-
-		assertTrue(selenium.isVisible("//td[1]/a"));
-		assertEquals(RuntimeVariables.replace("WC Structure1 Name"),
-			selenium.getText("//td[2]/a"));
-		assertEquals(RuntimeVariables.replace("WC Structure1 Description"),
-			selenium.getText("//td[3]/a"));
-		selenium.clickAt("//td[2]/a",
-			RuntimeVariables.replace("WC Structure1 Name"));
-		selenium.selectFrame("relative=top");
-
-		for (int second = 0;; second++) {
-			if (second >= 90) {
-				fail("timeout");
-			}
-
-			try {
-				if (selenium.isVisible("//iframe[@name='_15_changeStruture']")) {
-					break;
-				}
-			}
-			catch (Exception e) {
-			}
-
-			Thread.sleep(1000);
-		}
-
-		selenium.selectFrame("//iframe[@name='_15_changeStruture']");
 		selenium.type("//input[@id='_15_name_en_US']",
-			RuntimeVariables.replace("WC Structure2 Name"));
+			RuntimeVariables.replace("WC Structure1 Name"));
 		selenium.type("//textarea[@id='_15_description_en_US']",
-			RuntimeVariables.replace("WC Structure2 Description"));
+			RuntimeVariables.replace("WC Structure1 Description"));
 		selenium.clickAt("//input[@value='Add Row']",
 			RuntimeVariables.replace("Add Row"));
 		selenium.waitForPageToLoad("30000");
 		selenium.select("//select[@id='_15_structure_el0_type']",
 			RuntimeVariables.replace("Text Area (HTML)"));
 		selenium.type("//input[@id='_15_structure_el0_name']",
-			RuntimeVariables.replace("text 1"));
+			RuntimeVariables.replace("text"));
 		selenium.clickAt("//input[@value='Save']",
 			RuntimeVariables.replace("Save"));
 		selenium.waitForPageToLoad("30000");
 		assertEquals(RuntimeVariables.replace(
 				"Your request completed successfully."),
 			selenium.getText("//div[@class='portlet-msg-success']"));
-		assertTrue(selenium.isVisible("//tr[4]/td[1]/a"));
-		assertEquals(RuntimeVariables.replace("WC Structure2 Name"),
-			selenium.getText("//tr[4]/td[2]/a"));
-		assertEquals(RuntimeVariables.replace("WC Structure2 Description"),
-			selenium.getText("//tr[4]/td[3]/a"));
+		assertTrue(selenium.isVisible("//td[1]/a"));
+		assertEquals(RuntimeVariables.replace("WC Structure1 Name"),
+			selenium.getText("//td[2]/a"));
+		assertEquals(RuntimeVariables.replace("WC Structure1 Description"),
+			selenium.getText("//td[3]/a"));
 		selenium.selectFrame("relative=top");
 	}
 }
