+++ portal-web/test/com/liferay/portalweb/portlet/webcontentdisplay/wcstructure/addwcstructure2parentstructurestructure1wcd/RemoveParentStructureStructure1Structure2WCDTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portalweb.portlet.webcontentdisplay.webcontent.addwcstructure2parentstructure1wcd;
+package com.liferay.portalweb.portlet.webcontentdisplay.wcstructure.addwcstructure2parentstructurestructure1wcd;
 
 import com.liferay.portalweb.portal.BaseTestCase;
 import com.liferay.portalweb.portal.util.RuntimeVariables;
